Create Pace Creative British School
Set in the heart of Al Ajman we are a brand new school, offering affordable high quality education without compromise. We follow the UK British Curriculum through EYFS, Primary and Secondary, in time, culminating to GCSEs, BTECs and A Levels. Providing all our flourishing students the foundations necessary to be the change makers of tomorrow.
As part of the PACE Group we bring an extensive wealth of experience and knowledge spanning across 22 years, providing education from FS through to PHDs, across a variety of regions ,with a total student enrolment in excess of 23000. Education has been our legacy, our success and we will embed this at PACE Creative British School, Ajman.